requirements.txt
neo4j
redis
tensorflow-probability
tensorflow
pymc
numpy
pandas
scikit-learn
faiss-cpu
langchain
langchain-openai
langchain-community
streamlit
plotly
newspaper3k
feedparser
aiohttp
pydantic
apscheduler
gdelt
pymongo
requests
python-dotenv
arxiv
openai
beautifulsoup4
nltk
spacy
transformers
torch
sentence-transformers
uvicorn
fastapi